Update gcs.js to disable resumable uploads, this affects writes to version protected buckets
Writes to version protected buckets fail with a delete error on the file even if the user has admin permissions on the object/bucket. Not sure if it makes sense to have this as a flag, I'm sure the model would be to just disable resumable uploads becuase of the message oriented nature of the operations?